basic attributes are now a thing

+++ src/main.py
@@ -6,14 +6,24 @@ from jobTree.scriptTree.stack import Stack
 from jobTree.src.bioio import getLogLevelString, isNewer, logger, setLoggingFromOptions
 from lib.sqlite_lib import initializeTable
 from lib.general_lib import FileType, DirType, FullPaths
+
+#basic_attributes has many basic classes stuck together that are not really classifiers
+from src.basic_attributes import *
+
 from src.unknown_bases import UnknownBases
 
 #classifiers we are currently working with
 classifiers = [UnknownBases]
 
+#add in all of the basic attribute columns
+classifiers = classifiers + [TranscriptID, GeneID, GeneName, GeneType, TranscriptType]
+
+
 #hard coded file extension types that we are looking for
 alignment_ext = ".chained.psl"
 sequence_ext = ".fa"
+gene_check_ext = ".coding.gene-check.bed"
+#gene_check_details_ext = ".coding-gene-check-details.bed"
 
 def build_parser():
     """
@@ -25,6 +35,8 @@ def build_parser():
     parser.add_argument('--annotationBed', type=FileType)
     parser.add_argument('--alignmentDir', type=DirType, action=FullPaths)
     parser.add_argument('--sequenceDir', type=DirType, action=FullPaths)
+    parser.add_argument('--geneCheckDir', type=DirType, action=FullPaths)
+    parser.add_argument('--gencodeAttributeMap', type=FileType)
     parser.add_argument('--outDb', type=str, default="results.db")
     parser.add_argument('--primaryKey', type=str, default="AlignmentID")
     parser.add_argument('--overwriteDb', action="store_true")
@@ -41,17 +53,19 @@ def parse_dir(genomes, targetDir, ext):
     for g in genomes:
         path = os.path.join(targetDir, g + ext)
         if not os.path.exists(path):
-            raise RuntimeError("{} does not exist in {}".format(f, targetDir))
+            raise RuntimeError("{} does not exist".format(path))
         pathDict[g] = path
     return pathDict
 
 
-def build_analysis(target, alnPslDict, seqFastaDict, genomes, annotationBed, outDb, primaryKey, refGenome):
+def build_analysis(target, alnPslDict, seqFastaDict, geneCheckBedDict, gencodeAttributeMap,
+            genomes, annotationBed, outDb, primaryKey, refGenome):
     for genome in genomes:
         initialize_sql_table(genome, outDb, primaryKey)
         aln, seq = alnPslDict[genome], seqFastaDict[genome]
+        bed = geneCheckBedDict[genome]
         for classifier in classifiers:
-            target.addChildTarget(classifier(genome, aln, seq,
+            target.addChildTarget(classifier(genome, aln, seq, bed, gencodeAttributeMap,
                     annotationBed, outDb, refGenome, primaryKey))
 
 
@@ -68,24 +82,26 @@ def main():
     args = parser.parse_args()
     setLoggingFromOptions(args)
 
-    if args.overwriteDb is True:
+    if args.overwriteDb is True and os.path.exists(args.outDb):
         os.remove(args.outDb)
 
     logger.info("Building paths to the required files")
     alnPslDict = parse_dir(args.genomes, args.alignmentDir, alignment_ext)
     seqFastaDict = parse_dir(args.genomes, args.sequenceDir, sequence_ext)
+    geneCheckBedDict = parse_dir(args.genomes, args.geneCheckDir, gene_check_ext)
+    #geneCheckBedDetailsDict = parse_dir(args.genomes, args.geneCheckDir, gene_check_details_ext)
 
     refSequence = os.path.join(args.sequenceDir, args.refGenome + ".fa")
     if not os.path.exists(refSequence):
         raise RuntimeError("Reference genome fasta not present at {}".format(refSequence))
     args.refSequence = refSequence
 
-    i = Stack(Target.makeTargetFn(build_analysis, args=(alnPslDict, seqFastaDict, args.genomes, 
-            args.annotationBed, args.outDb, args.primaryKey, args.refGenome))).startJobTree(args)
+    i = Stack(Target.makeTargetFn(build_analysis, args=(alnPslDict, seqFastaDict, geneCheckBedDict, 
+            args.gencodeAttributeMap, args.genomes, args.annotationBed, args.outDb, args.primaryKey, 
+            args.refGenome))).startJobTree(args)
 
     if i != 0:
         raise RuntimeError("Got failed jobs")
-
 
 
 if __name__ == '__main__':
